extract reusable View class from UserForm
