+++ web/src/views/UserFrom.ts
@@ -1,21 +1,17 @@
+import { UserProps } from '../interfaces'
 import { User } from '../models/User'
+import { View } from './View'
 
-export class UserForm {
+export class UserForm extends View<UserProps> {
   constructor(
     public parent: Element,
     public model: User
   ) {
-    this.bindModel()
-  }
-
-  bindModel = (): void => {
-    this.model.on('change', () => {
-      this.render()
-    })
+    super(parent, model)
   }
 
   // mappping like react onClick function
-  private eventsMap(): { [key:string]: () => void } {
+  eventsMap(): { [key:string]: () => void } {
     return {
       'click:.button-set-name': this.onSetNameClick,
       'click:.button-set-age': this.onSetAgeClick
@@ -35,7 +31,7 @@ export class UserForm {
     this.model.setRandomAge()
   }
 
-  private template(): string {
+  template(): string {
     return `
       <div>
         <h1>User Form</h1>
@@ -50,32 +46,4 @@ export class UserForm {
     `
   }
 
-  // find all element inside in fragemnt and mapping events.
-  private bindEvents(fragment: DocumentFragment): void {
-    const eventsMap = this.eventsMap()
-
-    for (let eventKey in eventsMap) {
-      const [eventName, selector] = eventKey.split(':')
-
-      fragment.querySelectorAll(selector)
-        .forEach((element: Element) => {
-          element.addEventListener(eventName, eventsMap[eventKey])
-        })
-    }
-  }
-
-  // parse string to HTML and insert event
-  public render(): void {
-    // empty parent element
-    this.parent.innerHTML = '';
-
-    // create html element and rendering template
-    const templateElement = document.createElement('template')
-    templateElement.innerHTML = this.template()
-
-    this.bindEvents(templateElement.content)
-
-    // append to parent to show
-    this.parent.append(templateElement.content)
-  }
 }
